+++ plugin/src/main/java/nl/esciencecenter/e3dchem/knime/silicosit/preferences/PreferenceConstants.java
@@ -1,0 +1,13 @@
+package nl.esciencecenter.e3dchem.knime.silicosit.preferences;
+
+/**
+ * Constant definitions for plug-in preferences
+ */
+public class PreferenceConstants {
+
+	public static final String BABEL_LIBDIR = "BABEL_LIBDIR";
+
+	public static final String BABEL_DATADIR = "BABEL_DATADIR";
+
+	public static final String SILICOSIT_BINDIR = "SILICOSIT_BINDIR";
+}